pyproject.toml
[build-system]
requires      = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name            = "buttonbox"
version         = "0.1.0"
description     = "Configurable PC-side controller for a Raspberry Pi Pico button box"
requires-python = ">=3.9"
dependencies    = [
    "pyserial>=3.5",
    "pynput>=1.7",
    "click>=8.1",
    "PyQt6>=6.4",
    "evdev>=1.6; sys_platform == 'linux'",
]

[project.optional-dependencies]
gui = ["PyQt6>=6.4"]

[project.scripts]
buttonbox     = "buttonbox.cli:cli"
buttonbox-gui = "buttonbox.gui.__main__:main"

[tool.setuptools.packages.find]
where   = ["."]
include = ["buttonbox*"]
